+++ packages/client/src/events/participant.ts
@@ -1,33 +1,40 @@
-import { Dispatcher } from '../rtc';
-import { VisibilityState } from '../types';
+import { SfuEvent } from '../gen/video/sfu/event/events';
+import { StreamVideoParticipant, VisibilityState } from '../types';
 import { CallState } from '../store';
 
 /**
  * An event responder which handles the `participantJoined` event.
  */
-export const watchParticipantJoined = (
-  dispatcher: Dispatcher,
-  state: CallState,
-) => {
-  return dispatcher.on('participantJoined', (e) => {
+export const watchParticipantJoined = (state: CallState) => {
+  return function onParticipantJoined(e: SfuEvent) {
     if (e.eventPayload.oneofKind !== 'participantJoined') return;
     const { participant } = e.eventPayload.participantJoined;
     if (!participant) return;
-    state.setParticipants((participants) => [
-      ...participants,
-      { ...participant, viewportVisibilityState: VisibilityState.UNKNOWN },
-    ]);
-  });
+    // `state.updateOrAddParticipant` acts as a safeguard against
+    // potential duplicate events from the SFU.
+    //
+    // Although the SFU should not send duplicate events, we have seen
+    // some race conditions in the past during the `join-flow` where
+    // the SFU would send participant info as part of the `join`
+    // response and then follow up with a `participantJoined` event for
+    // already announced participants.
+    state.updateOrAddParticipant(
+      participant.sessionId,
+      Object.assign<StreamVideoParticipant, Partial<StreamVideoParticipant>>(
+        participant,
+        {
+          viewportVisibilityState: VisibilityState.UNKNOWN,
+        },
+      ),
+    );
+  };
 };
 
 /**
  * An event responder which handles the `participantLeft` event.
  */
-export const watchParticipantLeft = (
-  dispatcher: Dispatcher,
-  state: CallState,
-) => {
-  return dispatcher.on('participantLeft', (e) => {
+export const watchParticipantLeft = (state: CallState) => {
+  return function onParticipantLeft(e: SfuEvent) {
     if (e.eventPayload.oneofKind !== 'participantLeft') return;
     const { participant } = e.eventPayload.participantLeft;
     if (!participant) return;
@@ -35,18 +42,15 @@ export const watchParticipantLeft = (
     state.setParticipants((participants) =>
       participants.filter((p) => p.sessionId !== participant.sessionId),
     );
-  });
+  };
 };
 
 /**
  * An event responder which handles the `trackPublished` event.
  * The SFU will send this event when a participant publishes a track.
  */
-export const watchTrackPublished = (
-  dispatcher: Dispatcher,
-  state: CallState,
-) => {
-  return dispatcher.on('trackPublished', (e) => {
+export const watchTrackPublished = (state: CallState) => {
+  return function onTrackPublished(e: SfuEvent) {
     if (e.eventPayload.oneofKind !== 'trackPublished') return;
     const {
       trackPublished: { type, sessionId, participant },
@@ -57,24 +61,21 @@ export const watchTrackPublished = (
     // events, and instead, it would only provide the participant's information
     // once they start publishing a track.
     if (participant) {
-      state.updateOrAddParticipant(participant.sessionId, participant);
+      state.updateOrAddParticipant(sessionId, participant);
     } else {
       state.updateParticipant(sessionId, (p) => ({
         publishedTracks: [...p.publishedTracks, type].filter(unique),
       }));
     }
-  });
+  };
 };
 
 /**
  * An event responder which handles the `trackUnpublished` event.
  * The SFU will send this event when a participant unpublishes a track.
  */
-export const watchTrackUnpublished = (
-  dispatcher: Dispatcher,
-  state: CallState,
-) => {
-  return dispatcher.on('trackUnpublished', (e) => {
+export const watchTrackUnpublished = (state: CallState) => {
+  return function onTrackUnpublished(e: SfuEvent) {
     if (e.eventPayload.oneofKind !== 'trackUnpublished') return;
     const {
       trackUnpublished: { type, sessionId, participant },
@@ -82,13 +83,13 @@ export const watchTrackUnpublished = (
 
     // An optimization for large calls. See `watchTrackPublished`.
     if (participant) {
-      state.updateOrAddParticipant(participant.sessionId, participant);
+      state.updateOrAddParticipant(sessionId, participant);
     } else {
       state.updateParticipant(sessionId, (p) => ({
         publishedTracks: p.publishedTracks.filter((t) => t !== type),
       }));
     }
-  });
+  };
 };
 
 const unique = <T>(v: T, i: number, arr: T[]) => arr.indexOf(v) === i;