+++ packages/react-native-sdk/src/providers/StreamCall.tsx
@@ -11,7 +11,7 @@ import { MediaStreamManagement } from './MediaStreamManagement';
 
 export type StreamCallProps = {
   call: Call;
-  callCycleHandlers: CallCycleHandlersType;
+  callCycleHandlers?: CallCycleHandlersType;
 };
 /**
  * StreamCall is a wrapper component that orchestrates the call life cycle logic and