+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.33](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.0.32...@stream-io/video-react-bindings-0.0.33) (2023-07-03)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.0.1`
+### [0.0.32](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.0.31...@stream-io/video-react-bindings-0.0.32) (2023-07-03)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.0.1`
 ### [0.0.31](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.0.30...@stream-io/video-react-bindings-0.0.31) (2023-07-03)
 
 ### Dependency Updates